+++ tabletest-junit/src/test/java/io/github/nchaugen/tabletest/renderer/MarkdownRendererTest.java
@@ -3,6 +3,7 @@ package io.github.nchaugen.tabletest.renderer;
 import io.github.nchaugen.tabletest.parser.TableParser;
 import org.junit.jupiter.api.Test;
 
+import static io.github.nchaugen.tabletest.renderer.ColumnRoles.NO_ROLES;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class MarkdownRendererTest {
@@ -17,10 +18,13 @@ public class MarkdownRendererTest {
                 | --- | --- | --- |
                 | 1 | 2 | 3 |
                 """,
-            renderer.render(TableParser.parse("""
+            renderer.render(
+                TableParser.parse("""
                 a | b | c
                 1 | 2 | 3
-                """))
+                """),
+                NO_ROLES
+            )
         );
     }
 
@@ -32,10 +36,13 @@ public class MarkdownRendererTest {
                 | --- | --- | --- |
                 | 1 |  | 3 |
                 """,
-            renderer.render(TableParser.parse("""
+            renderer.render(
+                TableParser.parse("""
                 a | b | c
                 1 |   | 3
-                """))
+                """),
+                NO_ROLES
+            )
         );
     }
 
@@ -47,10 +54,13 @@ public class MarkdownRendererTest {
                 | --- | --- | --- |
                 | \\| | \\| | Text with \\| character |
                 """,
-            renderer.render(TableParser.parse("""
+            renderer.render(
+                TableParser.parse("""
                 a   | b   | 'a|b'
                 "|" | '|' | "Text with | character"
-                """))
+                """),
+                NO_ROLES
+            )
         );
     }
 
@@ -62,10 +72,13 @@ public class MarkdownRendererTest {
                 | --- | --- | --- |
                 | [] | [1, 2, 3] | [\\|, \\|] |
                 """,
-            renderer.render(TableParser.parse("""
+            renderer.render(
+                TableParser.parse("""
                 a  | b         | c
                 [] | [1,2,3] | ['|', "|"]
-                """))
+                """),
+                NO_ROLES
+            )
         );
     }
 
@@ -77,10 +90,13 @@ public class MarkdownRendererTest {
                 | --- | --- | --- |
                 | {} | {1, 2, 3} | {\\|\\|} |
                 """,
-            renderer.render(TableParser.parse("""
+            renderer.render(
+                TableParser.parse("""
                 a  | b   | c
                 {} | {1,2,3} | {"||"}
-                """))
+                """),
+                NO_ROLES
+            )
         );
     }
 
@@ -92,10 +108,13 @@ public class MarkdownRendererTest {
                 | --- | --- | --- |
                 | [:] | [a: 1, b: 2, c: 3] | [b: \\|\\|] |
                 """,
-            renderer.render(TableParser.parse("""
+            renderer.render(
+                TableParser.parse("""
                 a   | b             | c
                 [:] | [a:1,b:2,c:3] | [b: "||"]
-                """))
+                """),
+                NO_ROLES
+            )
         );
     }
 
@@ -107,10 +126,13 @@ public class MarkdownRendererTest {
                 | --- | --- | --- |
                 | [{}] | [[1], [2, 3]] | [[a: \\|], [b: [\\|]]] |
                 """,
-            renderer.render(TableParser.parse("""
+            renderer.render(
+                TableParser.parse("""
                 a    | b           | c
                 [{}] | [[1],[2,3]] | [[a: '|'], [b: ["|"]]]
-                """))
+                """),
+                NO_ROLES
+            )
         );
     }
 }