+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## [Unreleased]
 ### Added
-- Adding custom AsciiDoc role `expectation` to cells in columns where header name ends in `?`
+- AsciiDoc custom role `scenario` added to cells in scenario column
+- AsciiDoc custom role `expectation` added to cells in columns where header name ends in `?`
 ### Fixed
 - Preventing table values from being misinterpreted as AsciiDoc markup when rendered